Serve static assets from the app itself
Since they will be embedded in the app image, we want the app to always
serve these, not the proxy.

Static assets are moving from this model:

                                                    serve static asset
                                                    ^
                                                    |
    compiled assets volume - (read static asset) -> proxy (nginx)

To this model:

    serve static asset
    ^
    |
    proxy (nginx)
    ^
    |
    (proxy request)
    |
    app (embeds static assets)

We are making the app available on http://localhost:4000 so that devs
can keep the same development workflow.

Proxy is still available on http://localhost, even though static assets
are not being proxied correctly due to Access-Control-Allow-Origin.
Planning on addressing it via
https://medium.com/@yagoazedias_75857/how-to-configure-cors-in-your-phoenix-application-5ef0234bc25f

[#158222089]