Mitigate against make contrib failing when app deps change
We want to fail make contrib if exit code is not 2 (SIGINT - "Interrupt
from keyboard"), which is what happens on CTRL+C. This is the only non-0
exit code that is valid happy-path, otherwise we assume that
docker-compose up failed due to app dependencies.

The specific scenario which this fix mitigates against: if there is an
existing app image which has different dependencies baked in than what
is defined in the mix files, contrib will fail. The cached app image
will be re-created when build gets called since the layers have changed.

To always perform a clean container shutdown, we explicitly run
docker-compose down on SIGINT.

[#158222089]