Add make target for listing secrets stored in LastPass
This introduces a new make target, list-lp-secrets. Its purpose
is to list the majority of secrets stored in LastPass and used by this
app, as well as https://github.com/thechangelog/infrastructure, specifically
https://github.com/thechangelog/infrastructure/blob/master/.envrc

This commit removes .envrc, which is now redundant. All secret
environment variables are now captured in individual make targets.  These
targets are considered part of the private interface and are therefore
not exposed in `make help`.  For example, to load GitHub secrets in the
shell context, run `eval $(make github)`.  To load all secrets in the
shell context, run `eval $(make list-lp-secrets)`.

    Remember `gmake` vs `make` gotcha on OS X:
    https://github.com/thechangelog/changelog.com/commit/86891f4b74221cd0327301e46e674dcb190394be

The only .envrc functionality that was lost is `PATH_add script`.  This
was basically adding $PWD/script to $PATH and enabling users to run any
script from the `script` dir by invoking it directly: e.g. `run` vs
`script/run`.  Since we've introduced a Docker-based workflow in #246,
these scripts are less useful and I imagine them disappearing completely
in the not too distant future.  We might want to discuss this
some more with @jerodsanto before I take this idea to completion.

This commit also removes .envrc.personal, which is also redundant.
@jerodsanto picked up on the disconnect of requiring 2 separate .envrcs
instantly, so this commit solves the conflict by removing them
altogether.  The reason for needing 2 separate .envrcs is user-specific
secrets.  For example, CIRCLE_TOKEN is a personal API token for CircleCI.
I have my own, Jerod has his own, etc.  We can't reference a single
LastPass entry, since that would point to a specific token, which would
be the wrong one for either myself or Jerod.  Yes, we could use $USER to
distinguish between users, but it adds complexity, leaks knowledge
across boundaries, feels too smart for the value that it adds.  The
alternative that was chosen is to not version control .envrc and instead
provide a template that anyone can use as a starting point for their own
individual .envrc.  I am always open to better alternatives.

[finishes #161188291]