Add make target for running tests like CircleCI
CircleCI runs tests in a container, based on thechangelog/runtime image,
and uses the circleci/postgres:9.6 image for the db container. We need
to be able to do the same locally so that if CircleCI is not available
(regardless of the reason), changelog.com can carry on. Yes, things will
be more inconvenient (no auto-triggers), but the world won't stop for
changelog.com if CircleCI goes on holiday.

The end-goal is to be able to push changelog.com updates from local, old
skool style. To be at ease with this approach, we must be able to go
through the same process as CircleCI, including running tests.

[finishes #161685385]