Building a new proxy image will include the legacy assets
This is meant to address a potential WTF moment if anyone was to build
thechangelog/proxy image on their host without having the legacy static
assets present. The goal is for the legacy assets to be seamlessly
pulled just-in-time, at the time of building the proxy image, without
any extra hoops or WTF moments.

This approach makes use of Docker multi-stage builds:
https://docs.docker.com/develop/develop-images/multistage-build/

The resulting legacy_assets image is available via
https://hub.docker.com/r/thechangelog/legacy_assets/

[finishes #162167357]