Prevent incompatible dependencies reaching the Docker image
If cmark has been compiled on the host before running build or contrib,
and the platform is non-Linux (e.g. Darwin or Windows), the image will
fail to build. The reason being that deps gets copied into the image -
as is required when building images for production - but the resulting
NIF is platform-specific. Mounting a volume might shadow the image's
deps dir, but it's best to explicitly not copy this between host and
Docker image when running locally, in dev mode.

The downside of this is that an extra command needs to be run by those
that don't (can't) use make and need to run docker-compose manually.
Don't think it's worth spending more time over this, it's good enough
for now.

cc @jerodsanto this is likely to be the reason for the cmark-related
failures that you were getting when running the contrib target

[#162167357]