Get & compile deps when app starts in Docker
If there are local mix.exs changes, deps will need to be fetched &
compiled since the app was built. The build command acts like a cache,
meaning that the Docker image will have a fairly recent version of
dependencies already resolved & compiled. Fetching deps and compiling
everything when the app boots will most likely be a noop, but in some
rare cases it might resolve dependencies just-in-time.

Notice that we mount mix.exs & mix.lock. Any changes to these files made
on the host should be reflected in the running container & vice-versae.