Automatically mount db & uploads volumes to 2019 Linode instance
In hindsight, it would have made more sense to link the volumes to the
instance via the `config` block of the `linode_instance` resource. Since
the volumes were created after the instance was created, it felt more
natural to reference the instance id in the volumes. When we transition
to K8S, I expect all volume management to be handled by K8S, so this
solution has a short shelf-life, not much value in making it better now.

It's weird that we need to write the mount files to /tmp & then move
them to /etc/systemd/system, but we are not the only ones to hit this,
as captured in hashicorp/terraform#492. It's OK though, I expect K8S to
make this redundant.

It would be nice if the linode_image data source exposed a
default_ssh_user, as mentioned in
terraform-providers/terraform-provider-linode#21

[#162511937]