Change how the update container gets updated
Stop the current one first before starting a new one - we want a
single instance of the update container running. This is the default
behaviour, but I wanted to be explicit.

Wait 60 seconds before considering the update a success. Rollback if new
container fails within the first 60s.

Always restart the update container.

I am not clear on the behaviour when the container that is running the
docker stack deploy goes away. I am assuming that the Docker daemon
receives all instructions and knows what to do, even if the container
that started the docker stack deploy goes away. I am also assuming that
the new docker stack deploy, which runs in the new container, will
continue where the previous container left off. A lot of assumptions, I
know, but I am not aware of any documentation that explains what happens
in this case. Maybe @tianon can help... or at least he knows someone
that can help. Either way, willing to see this crash and burn, because
it's going to be amazing if will work as assumed ; )

[#131986287]