Limit continuous updates to the app service
We noticed that the db & proxy services get updated for no good reason,
especially when the image registry (Docker Hub) returns non-200 errors.
If we had more than 1 replica from each service type running on more
than 1 node, it wouldn't be a problem, but since there is only 1
instance running, every update results in downtime. The proxy service is
stateless and would be easy to change, but the db service requires
extensive modifications to run as a cluster. This is definitely the
easiest thing to do right now.