Disable CoreOS auto-updates
After the 1967.3.0 update, Docker failed to restart services. Previous
OS reboots following an auto-update resulted in ~20 minutes worth of
downtime.

Until we switch to a cluster of CoreOS hosts (or, even better, K8S),
and/or run the entire website via the CDN and serve a snapshot while
services are getting restored, let's disable these and control them
manually.

[#163178090]