Separate Docker Stack deployment from service updates in bootstrap image
This was creating issues when the Docker Stack was being updated via the
bootstrap container, and the update service would start by updating the
Docker Stack. I was always aprehensive about the dual nature of the
bootstrap script, but I kept ignoring the obvious refactoring until the
failures that I was running into became too annoying.

The nice side-effect of this refactoring is that now I can run the
updater locally as well, without worrying that it will interfere with
the stack when the update service gets updated.

When the stack/service are being updated, app_updater service needs to
be stopped, so that docker service/stack updates won't fail due to out
of sequence errors:

    failed to update service 2019_app: Error response from daemon: rpc error: code = Unknown desc = update out of sequence

[#163178093]