Try to update app service more frequently
App deploys in the new infrastructure take a few minutes longer. It's
mostly due to pushing/pulling images to/from Docker Hub. We might
want to consider a different Docker Image registry, and we might want to
consider web hooks to trigger service updates, but for now increasing
the image check frequency is the simplest thing.