Use 2048 bit Diffie-Hellman (DH) key exchange parameters
This improves SSL Labs score from B to A:
https://www.ssllabs.com/ssltest/analyze.html?d=2019.changelog.com

To make it an A+, as is the case of the current changelog.com, we would
need to remove all weak ciphers, specifically TLS_RSA_WITH_AES_* at the
Linode Node Balancer tier first, maybe via a "strong" cipher suite.
Currently only "weak" & "recommended" are the recognized cipher suites,
as described in the official Node Balancer Reference Guide:
https://www.linode.com/docs/platform/nodebalancer/nodebalancer-reference-guide/#tls-cipher-suites

Will open this discussion on the Linode Forum. Let me know @displague if
you have a better suggestion.

For extra context, this is the cipher suite configuration that achieves
the current A+ SSL Labs rating for changelog.com:
https://github.com/thechangelog/infrastructure/blob/0fb90109da8ce6f6f263df6b00aaeae56b74e51c/roles/nginx/files/ssl.http#L3

[finishes #163300195]