Add watch make target to easily watch all containers
It would have been easier to just add a target that lists all containers
is a format that is watch-friendly & then just e.g. `watch gmake
containers`, but the default macOS watch is not what Linux users expect.
This way, the "correct" watch gets installed if required, and there is
less to type. If `gmake` is already aliased to `m`, it's just `m w`. I
know, my laziness has no end... ¯\_(ツ)_/¯