Increase app start time to 60s
It was observed that during deploys, when there are 2 app instances
running, a lot of CPU may be used which makes a new app take more than 2
minutes to be ready. While we address the root causes (adding new app
instance too early to the proxy, poor default db config, unrestricted
container resources, etc.), bumping the start period from 30s to 60s
should be sufficient to have more reliable deploys.

Since this feels more related to environment than code, moving it from
the Docker image into the Docker Stack. Also, this changed more often
than I was expecting. Lastly, Docker Healthchecks are not recognised in
Kubernetes, which we are hoping to move to not before long, so we might
as well move this concern out of the image now.