Replace nginx-proxy with nginx locally
nginx-proxy is not ideal for proxying to backends which have their
lifecycle managed by Docker. nginx-proxy detects the new app container
and starts sending traffic to it *before* the Docker service considers
the new instance as healthy and promotes to live. Using the vanilla
nginx image and proxying to `app`, which gets resolved by Docker's DNS,
is the right approach in this context.

To switch to the new thechangelog/proxy locally:
* delete all local thechangelog/proxy images
* build a new proxy image (`gmake bpi`)
* either `gmake contrib` or `gmake ddsl`

You should now be using the new nginx-based proxy image.

To run end-to-end tests against the local proxy, run:

    gmake -f Makefile.test proxy LOCAL=1
    Testing gerhard.changelog.com resolving to 127.0.0.1...
     ✓ fqdn http
     - fqdn https (skipped: HTTPS is not configured for 127.0.0.1)
     ✓ nginx default_server
     ✓ legacy assets
     ✓ posts redirects
     ✓ changelog.fm
     ✓ gotime.fm
     ✓ jsparty.fm
     ✓ rfc.fm
     ✓ nginx_status is forbidden

To run end-to-end tests against the live proxy, run:

    gmake -f Makefile.test proxy
    Testing changelog.com resolving to 69.164.223.133...
     ✓ fqdn http
     ✓ fqdn https
     ✓ nginx default_server
     ✓ legacy assets
     ✓ posts redirects
     ✓ changelog.fm
     ✓ gotime.fm
     ✓ jsparty.fm
     ✓ rfc.fm
     ✗ nginx_status is forbidden
       (in test file test/e2e/proxy.bats, line 61)
         `grep "403 Forbidden"' failed

nginx_status is failing on changelog.com because it is not currently
enabled. The new proxy has it enabled in preparation for netdata
integration.

[#164473414]