Use ip_address attribute to retrieve public IPv4 address
If there is a public & private IPv4 address, ipv4 on linode_instance
(v1.3.0) returns an array, which is not compatible with dnsimple_record
(v0.1.0) - this expects a string.  I wouldn't expect for the type of
ipv4 to change based on the state of linode_instance.

What do you think @displague?

re #283

[#161198623]