Add make target to help find flaky tests
I got 1 failure in 10, committing so that we can use it in determining
which tests are most flaky. Given sufficient runs, we should see same
test failures repeating. These are good candidates for fixing or
deleting.

@jerodsanto try 100 runs and commit what you find:

    gmake find-flaky-tests TEST_RUNS=100

Maybe there is a better way of doing this. In Erlang, Common Test has
this feature built-in:
https://learnyousomeerlang.com/common-test-for-uncommon-tests