Use explicit FROM image version for production
It would be too easy to bump runtime and base production on top of a
version that we are not ready for.

Thanks https://github.com/hadolint/hadolint!

FWIW, we should bump circleci/elixir:1.6-node to
circleci/elixir:1.8-node. Any reason why we shouldn't @jerodsanto?
It ships with:

    YARN_VERSION=1.13.0
    OTP_VERSION=21.3.8.1
    REBAR_VERSION=2.6.4
    ELIXIR_VERSION=v1.8.1
    NODE_VERSION=10.15.3
    REBAR3_VERSION=3.10.0

We are currently on:

    YARN_VERSION=1.9.4 # a couple minors behind
    OTP_VERSION=20.3.8.11 # a major behind
    REBAR_VERSION=2.6.4
    ELIXIR_VERSION=v1.6.6 # a few minors behind
    NODE_VERSION=8.12.0 # a few majors behind
    REBAR3_VERSION=3.6.1

BTW, OTP 22 landed this week: http://www.erlang.org/news/132