[admin] Add foreign key constraints to post form
Without this in place, if you manually fiddled around with the HTML
source of the input fields, you could input an invalid topic_id which
causes Ecto to throw an exception since it violates the constraint at
the database level.

This patch adds the validation at the changeset level so we can display
a friendly error if the FK ids aren't available.