Add /health route for doing health checks
Previously the health check was hitting the home page which rendered a
full response back of the home page and wrote a log entry for that
request.

With this new set up, an empty 200 gets returned and since it's loaded
at the beginning of the Endpoint, it bypasses the logger and other
plugs which helps eliminate log spam and serve the health check more
efficiently.