Disable netdata proxying
It couples nginx to netdata in an unhealthy way. Netdata expects nginx
to be running for it to start capturing metrics. If nginx depends on
netdata, and netdata - for whichever reason - goes away, any request to
netdata proxied via nginx will make nginx crash because the netdata host
won't be resolvable by DNS. This is bad on many levels, but the focus is
on the unhealthy coupling between these 2 services.

[#167116050]