Disable cgroups & statsd in netdata
For cgroups we want to be using cadvisor, netdata's integration with
Docker is sub-par. This is a follow-up from
https://github.com/thechangelog/changelog.com/commit/db6ea6d733bc0644bd304b5cc585cacc80eed14b,
which has more context.

Statsd is simply not used, no point in having it burn CPU cycles.

Also, use a local netdata config to test changes in local production
before taking them live.

Query netdata via the service name, this ensures that the DNS, and the
service networking is working correctly.

[#167116050]