Pin Terraform to v0.11
We were planning to upgrade to v0.12, but then we learned that the
biggest obstacle to this upgrade is the Linode host name, which contains
only digits: 2019. Apparently this is not allowed in Terrafrom v0.12,
meaning that we would need to migrate onto a new Linode instance in
order to upgrade to Terraform v0.12. I need to take a closer look at
this, but it's definitely something that I want to defer for the time
being.

[finishes #167116483]