Promote 201910 stack to live
  make ctop
  # stop db_backup container so that service restarts it and db gets backed up

  make -f Makefile.201910 dbri
  # restore from db backup triggered above

  make -f Makefile.201910 ru
  # one last uploads sync

  make iaas
  # promote 201910 stack to live

Notice that we have added the new host to the previous stack's
NodeBalancer so that we don't depend just on DNS propagation for the
switch-over. This way, the update is instant, and by the time we delete
all infrastructure that was running the previous stack, the DNS will
have propagated for everyone, so we will be able to delete the previous
stack's NodeBalancer with no service disruption.