Fix changelog ten Docker image assets
Because the upstream asset URLs change, the easiest thing is to go with
the single file & embedded assets approach. It's slower to load, but it
won't break if the asset URLs change as everything is embedded  in the
HTML.

monolith seemed like a good idea, but in practice the HTML would look
broken. While SingleFile breaks JS, everything looks the way it should.

Signed-off-by: Carmina Lazu <carmina@lazu.co.uk>