Add target for updating all LKE nodes to latest available K8S version
    export LKE_LABEL=prod-20200426
    make lke-update

The first implementation of pool-recycle is MVP, meaning that nodes get
updated all at once. Combined with persistent volume detach lag and
calico slow startup, this can result in ~5 minutes of downtime. A better
implementation would imply nodes cycling 1-by-1, and a faster volume
detach. While this command works, it's the quick and dirty equivalent
to:

    export LKE_LABEL=prod-20200426
    make lke-pool
    # wait for new nodes to be ready
    # drain & cordon old nodes, 1-by-1
    # delete old nodes, after all workloads have migrated to the new pool

[delivers #174067023]

Signed-off-by: Carmina Lazu <carmina@lazu.co.uk>