Serve legacy assets from the app itself
Rather than bundling these in the nginx proxy, let's bundle them into
the app instead. The goal is to remove the nginx proxy in K8S and just
use ingress-nginx (because 1 nginx is enough).

Notice the lowercase headers. They are case-insensitive in HTTP/1.1 and
lowercase in HTTP/2, which we are moving towards.

- HTTP/1.1 https://tools.ietf.org/html/rfc7230#section-3.2
- HTTP/2 https://tools.ietf.org/html/rfc7540#section-8.1.2

[finishes #174509798]

Signed-off-by: Carmina Lazu <carmina@lazu.co.uk>