Only the same tag will force app updates via keel
Because we push both $CIRCLE_SHA & latest tags, there will be 2 events
sent to keel.changelog.com, meaning 2 updates. As we only really care
about latest, this is the simplest solution. GitOps via ArgoCD is worth
exploring in the future, for sure. For now, this is closest to what we
currently have in the Docker Swarm world.

[#174509592]

Signed-off-by: Carmina Lazu <carmina@lazu.co.uk>