[admin] Fix bug where ‘all results’ search page raises error
The problem with `render(json_template, …` is it assumes everything we’re
rending is json, but this same view also renders `all.html` and there
is no way to pattern match that. 

Instead, we “moisten” the code and have an explicit `render` function 
for each json template we know about.