Deploy grafana-agent in the grafana-cloud namespace
Even though everything is up and running, and there are no errors in the
logs, the agent is not scraping any targets according to its
agent_prometheus_ha_configs_created_total metric.

Leaving this here as an intermediary step, but considering that the
agent is still in beta, configuring Prometheus' remote_write seems like
a more robust approach.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>