Deploy kube-prometheus-stack using Helm
* configure app to be scraped
* configure PromEx to setup dashboards in the local Grafana

Even though remote writing to Grafana Cloud works, We are not yet ready
to send 70k metrics to it. Waiting for @adamstac to confirm when we are
good to enable this.

As a short-term alternative, we may be able to configure Grafan Cloud to
connect to the Prometheus running in LKE, so that we can see all metrics
and logs in Grafana Cloud, even though we are not yet storing all
metrics there.

Had to delete the previously deployed kube-prometheus-stack because the
new one wouldn't work correctly due to CRD mismatches. Will restore
grafana.changelog.com as a next step, before deleting the config that we
were using previously.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>