Increase prod db connection pool from 20 to 40
Db operations can spend as long as 10s in the queue when select
statements take a few seconds instead of tens of milliseconds. While
this is unlikely to solve the underlying problem, it shouldn't harm.
We have all the metrics that we could possibly want & need to see how
well this works in practice. Shipped!

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>