Use PKCS8 encoding for *.changelog.com secret key
POST https://api.fastly.com/tls/private_keys will fail with 400
otherwise.

It was easier to request a new certificate, wildcard-acme-tls-pkcs8, and
update all references to it rather than forcing re-issuance of
wildcard-acme-tls. Had to manually delete the wildcard-acme-tls cert
object and the corresponding secret. It may seem fiddly, but it was
pretty straightforward in reality.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>