Get app version from env, default to fixed date
This version never changed since the app was created, and it should. As
we don't do semver (even though Elixir insists on it) we pick a semver
that makes sense: calver disguised as semver 😀

The end-result is a version that changes between builds, and a value in
PromEx that makes it useful to find out which build introduced a
variation of interest in the app metrics.

While at it, capture GIT_SHA & GIT_AUTHOR in the pipeline so that
PromEx has everything it needs to annotate properly.

If APP_VERSION variable is not set, default to what we've been always
using: 0.0.1

cc @jerodsanto @akoutmos in case you are interested in this change

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>