Do not fail on Fastly POST /tls/private_keys
It's not an idempotent operation, the request will fail with 400 if the
key exists. It would be great if we could check if the key exists first,
before POST-ing. It would be even better if we could PATCH instead of
POST, same as we do for the certificate, but the Fastly API doesn't
support that request type.

The logs were too verbose by default, especially with all the TLS info.
Only run curl in --verbose mode when the job itself runs in debug mode,
otherwise keep it on the down low, as the UNIX philosophy recommends.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>