Install OpenFaaS straight into K8S & arkade locally
https://twitter.com/alexellisuk/status/1347526129708625921

This was done part of a pilot podcast. We were chatting OpenFaaS with
@alexellis and then just wanted to try it out... straight into
production 😲

I know, some may think You're Crazy! I see it as owning production and
not fearing mistakes. Our setup is resilient, and I understand it inside
out, so the worst that can happen is learning about something that I
didn't know, and then just improve it.

And if you still think that doing it straight in production is scary and
crazy, think of this: we drive cars all the time, sometimes with our
entire family in it. If one is responsible enough to do that, then this
is low stakes in comparison.

You'll have to wait some number of weeks before the episode goes out,
and maybe up to a week for the video, but this is going straight into
the main branch, and into production 👍

A few things for when we continue with this:

- https://github.com/openfaas/faas-netes/blob/0.12.12/chart/openfaas/README.md

- We cannot expose secrets as envs to OpenFaaS functions, only as files:
  https://docs.openfaas.com/reference/secrets

- If we use the prod-2020-07 namespace for OpenFaas functions, where the
  changelog.com app & db runs, how do we make OpenFaaS not consider
  changelog.com app & db pods as functions?

Thanks @alexellis for the good fun!

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>