Migrate from Crunchy Data to Zalando PostgreSQL
Crunchy Data has not been the best experience. Based on the initial
setup, Zalando PostgreSQL seems to be better.

What worked well with Crunchy Data:

* it didn't lose any data
* it degraded gracefully under load

What didn't work so well:

* the initial setup was complicated, with too much documentation that
  was poorly structured (no teasing, lumps of unrelated items, probably
  useful, but confusing as hell)
* default backups filled the disk and made the db become unavailable
  (not a safe default & not easy to configure)
* the pgo client approach is not GitOps friendly, I really missed the
  CRD approach
* replication stopped, pg-wal on the primary filled up, and eventually
  the primary crashed and wouldn't restart - that was a fun one to debug
  & recover from (our own custom backups saved the day)
* pods have too many labels (15+) which means that Loki running in
  Grafana Cloud rejects the db logs as they breach the maximum number of
  allowed labels 🙄
* couldn't figure out how to get metrics into Prometheus & Grafana
  dashboards into our kube-prometheus-stack setup

All in all a poor experience, too heavyweight and complicated. But it's
OSS, so we tried it, captured this feedback, and moved on because YOLO.

In closing, within the first 30 minutes of switching to a PostgreSQL
provisioned by the Zalando Operator, our 95th responses went from 313ms
avg to 228ms according to nginx logs. That's a 28% speedup by simply
switching the operator. So far, so good 😉

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>