Use host from config for all redirects
Rather than hard-coding this to "changelog.com", we want to pick the
host from config, which is currently set via URL_HOST.

This used to be changelog.com, but we are moving to www.changelog.com.
This is required to configure a CNAME DNS record which points to
dualstack.changelog.map.fastly.net, our Fastly CDN service.

changelog.com is the apex and we cannot have a CNAME for it - DNS
limitation
https://docs.fastly.com/en/guides/using-fastly-with-apex-domains.
Rather than hard-coding multiple anycast IPs, and working around the
AAAA limitation in external-dns (currently not supported), the easiest
thing to do is to route all traffic through www.changelog.com, which is
already pointing to Fastly via CNAME. apple.com, google.com &
netflix.com do the same. github.com doesn't, but majority wins.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>