Enable ingress-nginx to serve multiple changelog.com subdomains
Had to configure external-dns to stop watching ingresses & services, and
only watch dnsendpoints. This way, ingress-nginx can service hosts
without worrying that external-dns will automatically want to manage
them.

While this is not a change that was technically needed, the relationship
between ingress-nginx & external-dns felt awkward and caused trouble in
the past when we add Fastly into the mix. This is better now as it's
more explicit and fits with the way we run ops. Also, it fits with our
2021 H2 plans. But one step at a time...

Next step: https://3.basecamp.com/4543941/buckets/19947330/todos/4014320897

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>