Skip grafana-agent for now in new 2022 LKE
The focus is on getting the entire MVP stack working, and grafana-agent
is not a hard requirement.

Also, stick to ingress-nginx & cert-manager instead of traefik. As much
as I want to try it out with the hope that it will simplify the stack,
the focus is on sticking to what we know and get this out there ASAP, so
that we can start testing @jerod's S3 integration.

Next step: cert-manager.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>