Run changelog.static.upload mix task before all other tasks
If we run it part of the same chain, it terminates it before phx.server
can run. Maybe @lawik or @akoutmos know why this happens. We look at
ecto.create and there is nothing special about it, so our task should
work the same, but for some reason it doesn't and we don't know why 🤷

Pair: @jerodsanto

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>