Deploy https://changelog-2022-03-13.fly.dev/
Most of this was done imperatively, by running fly CLI. Some of it was
captured in the main.cue. The focus of this is to get this working, have
a discussion, and then to make it right.

Capture a bunch more fly commands as dagger actions

Deploy to fly.io if prod_image gets built & pushed successfully

This *must* be moved into a standalone fly workflow right after we
migrate. While this was the simplest thing, it's wrong to combine build
& deploy concerns into a single pipeline.

This next step requires build_test_ship to commit a fly.toml update, and
this workflow triggering on fly.toml updates

🔥 This is fine for now 🔥

⚠️  cue.mod is for Dagger 0.2.x and will break the
0.1.x workflow that LKE 2022 current prod depends on. ⚠️

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>