Hard-code cookie domain to changelog-2022-03-13.fly.dev
This is a *temporary* change and *must* be reverted before this gets
merged, otherwise new admin logins on changelog.com will stop working.

This commit shows how to get admin logins working on fly.io, when the
app is running behind changelog-2022-03-13.fly.dev. Due to security
reasons, .fly.dev. domains are *not* allowed and will *not* work.

The gotcha is that this is a *compile* config, meaning that setting it
in the env is *not* going to work.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>