Upgrade Elixir from 1.12 to 1.13
Run `make howto-upgrade-elixir` and go through the steps. This is not as
accessible as it could be. Intend to move it into the README.md next.

While doing this, I bumped Erlang to latest 24, specifically 24.3.4.4. I
also took the opportunity to upgrade the container OS from Ubuntu Focal
(20.04 LTS) to Ubuntu Jammy (22.04 LTS). If we discover any issues with
ffmpeg or imagemagick, we may want to re-consider this bump.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>