Upgrade Elixir from 1.12 to 1.13 (#423)
* Upgrade Elixir from 1.12 to 1.13

Run `make howto-upgrade-elixir` and go through the steps. This is not as
accessible as it could be. Intend to move it into the README.md next.

While doing this, I bumped Erlang to latest 24, specifically 24.3.4.4. I
also took the opportunity to upgrade the container OS from Ubuntu Focal
(20.04 LTS) to Ubuntu Jammy (22.04 LTS). If we discover any issues with
ffmpeg or imagemagick, we may want to re-consider this bump.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Upgrade Node from 14.17 to 14.20

It felt a sensible thing to do.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Upgrade PostgreSQL from 12.6 to 14.1

That's what we use currently in production (Fly.io). I should have done
this a long time ago... Kaizen Driven Development FTW!

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Bump Elixir version to 1.13 in README & mix.exs

The app will now fail to boot if Elixir 1.13 (or newer) is not
available. It also hints to those that read the README what is the min.
required Elixir version.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Add how to upgrade Elixir steps to README

Remove them from the Makefile, we only want them in a single place.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>