Use image in devcontainer instead of Dockerfile
So that contributors are not to forced for everything to build from
scratch. We are also thinking of improving on those Dockerfiles so that
we can make better use of op caching. All of it can be captured as code
that will be used as one big CI/CD pipeline, with sweet caching.

This is a follow-up to https://github.com/thechangelog/changelog.com/pull/437

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>