Build & publish ghcr.io/thechangelog/changelog-runtime from GitHub Actions using Dagger - still works locally (#450)
* Build & publish runtime container image with Dagger

We used to do this locally, via a Makefile, and publish to
https://hub.docker.com/r/thechangelog/runtime/tags. This used to be a
manual process.

We are now publishing to ghcr.io/thechangelog/changelog-runtime using
Dagger v0.3 Go SDK. All relevant code is in magefiles/image. We can
still run this locally via:

	cd magefiles && go run main.go image:runtime

While the above is great for testing, we want our CI/CD doing this work.
As a result, this is now part of our `ship_it.yml` GitHub Actions
workflow. BTW, notice the new pipeline entrypoint: `go run main.go -w ../ ci`

This new entrypoint hints to the fact that we will use GitHub Actions to
trigger Dagger, which in turn will kick-off all the pipelines. That's
right. Because of the amazing built-in caching, we can run everything on
every commit, and only the parts that changed will re-run. If the
runtime image changes, everything else that depends on it will change
too. If it doesn't, the other pipelines will simply re-use the cached
layers. No more messing with paths in GitHub Actions. Dagger Engine will
figure it out. If you think that this is cool, wait when the rest of the
stuff that is currently in the works ships 🤯

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

---------

* Bump runtime image to ghcr.io/thechangelog/changelog-runtime

Built in GitHub Actions with Dagger v0.3 Go SDK

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>