Migrate Dagger v0.1 to Dagger Go SDK v0.5 (#454)
This is a follow-up to
https://github.com/thechangelog/changelog.com/pull/446

With service containers support - a.k.a. container-to-container
networking - landing in the recently released Dagger Go SDK v0.5, we
can now run a test PostgreSQL db straight in our Dagger pipeline, a `mix
test` requirement. This means that we can finally migrate from Dagger v0.1 💪

FTR: https://github.com/dagger/dagger/pull/4505

With this in place, we can also revert
https://github.com/thechangelog/changelog.com/commit/83e8268bdcccef4604716e0ced3fdf0fd5b2c0de

We are also uploading static assets at build time, not at boot time.
This will speed boot times considerably. We are spending more than 1
minute on every app start due to those assets being uploaded (most of
them are getting skipped anyways - an improvement for another day).

Notice the s3_static.ex change. We are using ExAws.request! so that the
task will fail if any of the static asset files cannot be uploaded to
S3. Without this change, ops can fail silently and we will not know
about it.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>