Upgrade runtime deps - April, 2023 session (#455)
* Update Dagger Go SDK to latest stable

Also panic if container images cannot be exported locally when
debugging. This surprised me. `tmp` did not exist and this was failing
silently.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Upgrade Node.js from 14 LTS to 18 LTS (latest)

This makes it obvious where we would need to template this version.

Thanks Jamie Tanna for the nudge!
https://changelog.slack.com/archives/C024Q4CER/p1680887511450349?thread_ts=1680875047.233929&cid=C024Q4CER

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Upgrade Erlang & Elixir to latest stable

This goes naturally with the Node.js upgrade.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Bump flyctl to latest stable

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Extract err handling in must* functions

Thanks Jamie!

While at it, refactor & remove the adjacent code.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

---------

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>