Remove the last Makefile
This was meant to be removed for Kaizen 9 -
https://github.com/thechangelog/changelog.com/discussions/440 - but I
was focusing on other things, and it slipped through.

Listening back to that episode - https://changelog.com/shipit/90 -
reminded me to check if we still have any Makefile left, so here it
goes: the end of Makefiles in changelog.com. Change is good!

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>