Change sign-in flow to require a POST instead
Some email clients (Outlook!) will visit all links prior to the user 
clicking on them. (This is probs a malware/phishing detecting measure.)

That sucks because it "uses" their sign-in token prior to them getting
to actually use it to sign in.

So, I'm testing a change to this flow where a GET request to that path
renders a hidden form that submits a POST via JS to do the actual signing
in part.

If JavaScript is disabled, the form shows a button to click to manually
sign yourself in. Seems like a decent