Read versions from .tool-versions when building production image
In order to minimize bugs and differences in behaviour between dev &
prod, we want to be running the same runtime dependency versions (e.g.
Erlang, Elixir, Node.js & PostgreSQL) in both environments.

In development, we are using https://asdf-vm.com/ to manage runtime
dependencies. This means that with a single command - asdf install -
everyone working with our codebase can have the same dependencies
installed. Currently these are:
- direnv 2.32.2
- elixir 1.14.4
- erlang 25.3.2
- golang 1.20.4
- nodejs 18.16.0
- yarn 1.22.19
- postgres 14.1
- flyctl 0.1.8
- mage 1.15.0

We used to manually edit these versions in our CI/CD pipeline. Since we
are using Dagger for CI/CD, this was no big deal since our pipeline is
just code. But since it is code, we could go one step further. Our
pipeline could read the .tool-versions file and use that information
to build the container image that gets deployed to production. That's
exactly what this change does.

The only runtime dependency version that does not exist in
.tool-versions is the OS - Ubuntu 22.04 in production. This is a
reasonable limitation that we could improve as a follow-up. What I am
thinking right now is that we should NOT depend on an OS for the
production container image, and instead finish the migration to Elixir
(Erlang) releases + one of those sweet
https://github.com/chainguard-images/images#chainguard-images, maybe
https://github.com/chainguard-images/images/tree/main/images/glibc-dynamic.
What we are really missing is an Elixir Chainguard image. One day...

Boy scout fixes:
+ upgrade Erlang, Node.js & Golang to latest minor
+ upgrade Dagger Go SDK to latest
+ go mod tidy
+ improve Dagger pipeline grouping
+ run pipelines in parallel 🏎 Turbo button 🏎
+ use the same PostgreSQL minor in devcontainer - same version everywhere!
+ upgrade flyctl version to latest
+ install yarn, flyctl & mage via asdf
+ use shorter mage commands instead of go run
+ test asdf on a fresh macOS 16.3 (arm64)
+ test asdf on Debian 11
+ upgrade setup-go GitHub Action to v4

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>